[Fix] 즐겨찾기된 아이템 Fetch 로직 수정

+++ Projects/Domain/Sources/Entity/MarketType.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-public enum MarketType: Int, CaseIterable {
+public enum MarketType: Int64, CaseIterable {
     case overseas, domestic
     
     public var toString: String {
